requirements.txt
opencv-python-headless 
deepface
pandas
pillow
numpy
tf-keras